début du sequelize

+++ backend/controllers/post.js
@@ -1,57 +1,29 @@
-const db = require("../db");
 const fs = require("fs");
+const { Post, User } = require("../models");
 
 exports.create = async (req, res) => {
   if (!req.body.title || !req.body.content)
     return res.status(400).json({ error: "Bad Request" });
   try {
-    await db.execute(
-      `
-      INSERT INTO posts
-      (title,content,author_id,media)
-      VALUES
-      (?,?,?,?)
-    `,
-      [
-        req.body.title,
-        req.body.content,
-        req.auth.userId,
-        req.file ? req.file.filename : null,
-      ]
-    );
+    await Post.create({
+      title: req.body.title,
+      content: req.body.content,
+      userId: req.auth.user.id,
+      media: req.file ? req.file.filename : null,
+    });
     res.status(201).json({ message: "Post Created" });
   } catch (e) {
     console.error(e);
-    if (req.file) {
-      fs.unlinkSync(req.file.path);
-    }
+    if (req.file) fs.unlinkSync(req.file.path);
     res.status(500).json({ error: "Internal Server Error" });
   }
 };
 
 exports.list = async (req, res) => {
   try {
-    const [posts] = await db.execute(
-      `
-      SELECT
-      p.*,
-      u.id AS author_id,
-      u.firstname AS author_firstname,
-      u.lastname AS author_lastname,
-      u.avatar AS author_avatar,
-      COUNT(l.user_id) AS likes,
-      EXISTS(SELECT * FROM likes l WHERE l.post_id = p.id AND l.user_id = ? LIMIT 1) AS is_liked
-      FROM posts p
-      JOIN users u 
-      ON p.author_id = u.id 
-      LEFT JOIN likes l
-      ON l.post_id = p.id
-      WHERE p.parent_id IS NULL
-      GROUP BY p.id
-      ORDER BY p.created_at DESC
-    `,
-      [req.auth.userId]
-    );
+    const posts = await Post.findAll({
+      include: [User],
+    });
     res.status(200).json(posts);
   } catch (e) {
     console.error(e);
@@ -61,29 +33,9 @@ exports.list = async (req, res) => {
 
 exports.get = async (req, res) => {
   try {
-    const [posts] = await db.execute(
-      `
-      SELECT
-      p.*,
-      u.id AS author_id,
-      u.firstname AS author_firstname,
-      u.lastname AS author_lastname,
-      u.avatar AS author_avatar,
-      COUNT(l.user_id) AS likes,
-      EXISTS(SELECT * FROM likes l WHERE l.post_id = p.id AND l.user_id = ? LIMIT 1) AS is_liked
-      FROM posts p
-      JOIN users u 
-      ON p.author_id = u.id 
-      LEFT JOIN likes l
-      ON l.post_id = p.id
-      WHERE p.id = ? AND p.parent_id IS NULL
-      GROUP BY p.id
-      
-    `,
-      [req.auth.userId, req.params.id]
-    );
-    if (!posts.length) return res.status(404).json({ error: "Post Not Found" });
-    res.status(200).json(posts[0]);
+    const post = await Post.findByPk(req.params.id, { include: ["comments"] });
+    if (!post) return res.status(404).json({ error: "Post Not Found" });
+    res.status(200).json(post);
   } catch (e) {
     console.error(e);
     res.status(500).json({ error: "Internal Server Error" });
@@ -94,12 +46,15 @@ exports.modify = async (req, res) => {
   if (!req.body.title || !req.body.content)
     return res.status(400).json({ error: "Bad Request" });
   try {
-    const [result] = await db.execute(
-      "UPDATE posts SET title = ?, content = ? WHERE id = ? LIMIT 1",
-      [req.body.title, req.body.content, new Date(), req.params.id]
+    const array = await Post.update(
+      {
+        title: req.body.title,
+        content: req.body.content,
+      },
+      { where: { id: req.params.id } }
     );
-    if (!result.affectedRows)
-      return res.status(404).json({ error: "Post Not Found" });
+    console.log(array);
+    if (!array.length) return res.status(404).json({ error: "Post Not Found" });
     res.status(200).json({ message: "Post Modified" });
   } catch (e) {
     console.error(e);
@@ -109,12 +64,8 @@ exports.modify = async (req, res) => {
 
 exports.delete = async (req, res) => {
   try {
-    const [result] = await db.execute(
-      "DELETE FROM posts WHERE id = ? LIMIT 1",
-      [req.params.id]
-    );
-    if (!result.affectedRows)
-      return res.status(404).json({ error: "Post Not Found" });
+    const array = await Post.destroy({ where: { id: req.params.id } });
+    if (!array.length) return res.status(404).json({ error: "Post Not Found" });
     res.status(200).json({ message: "Post Deleted" });
   } catch (e) {
     console.error(e);
@@ -123,8 +74,6 @@ exports.delete = async (req, res) => {
 };
 
 exports.like = async (req, res) => {
-  // res.status(200).json( req.body );
-  // return;
   try {
     if (req.body.isLiked) {
       await db.execute(
@@ -134,7 +83,7 @@ exports.like = async (req, res) => {
         VALUES
         (?,?)
       `,
-        [req.params.id, req.auth.userId]
+        [req.params.id, req.auth.user.id]
       );
       res.status(200).json({ message: "Post Liked" });
     } else {
@@ -142,7 +91,7 @@ exports.like = async (req, res) => {
         `DELETE FROM likes
          WHERE post_id = ? AND user_id = ?
          LIMIT 1`,
-        [req.params.id, req.auth.userId]
+        [req.params.id, req.auth.user.id]
       );
       res.status(200).json({ message: "Post Unliked" });
     }
