+++ backend/controllers/user.js
@@ -1,14 +1,10 @@
 /**
  * importation bibliothèques
  */
-const jwt = require("jsonwebtoken");
-const bcrypt = require("bcrypt");
 const passwordValidator = require("password-validator");
-const db = require("../db");
-const env = require("../env");
+const { User } = require("../models");
 const fs = require("fs");
 const path = require("path");
-const isModerator = require("../middleware/moderator")
 const regexEmail =
   /(?:[a-z0-9!#$%&'*+/=?^_`{|}~-]+(?:\.[a-z0-9!#$%&'*+/=?^_`{|}~-]+)*|"(?:[\x01-\x08\x0b\x0c\x0e-\x1f\x21\x23-\x5b\x5d-\x7f]|\\[\x01-\x09\x0b\x0c\x0e-\x7f])*")@(?:(?:[a-z0-9](?:[a-z0-9-]*[a-z0-9])?\.)+[a-z0-9](?:[a-z0-9-]*[a-z0-9])?|\[(?:(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)\.){3}(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?|[a-z0-9-]*[a-z0-9]:(?:[\x01-\x08\x0b\x0c\x0e-\x1f\x21-\x5a\x53-\x7f]|\\[\x01-\x09\x0b\x0c\x0e-\x7f])+)\])/;
 
@@ -38,10 +34,12 @@ exports.signup = async (req, res) => {
     !req.body.password
   )
     return res.status(400).json({ error: "Bad Request" });
+  const { firstname, lastname, email, password } = req.body;
+
   /**
    * condition si email non validé retourne erreur 400
    */
-  if (!regexEmail.test(req.body.email))
+  if (!regexEmail.test(email))
     return res.status(400).json({ error: "Invalid Email Address" });
 
   /**
@@ -49,23 +47,20 @@ exports.signup = async (req, res) => {
    * retourne un tableau d'erreur sinon vide = true
    * tableau vide = mot de passe valide
    */
-  const details = passValidator.validate(req.body.password, { details: true });
+  const details = passValidator.validate(password, { details: true });
   if (details.length)
     return res
       .status(400)
       .json({ error: `Invalid Password: ${details[0].message}` });
 
   try {
-    /**
-     * hash du mot de passe
-     */
-    const hash = await bcrypt.hash(req.body.password, 10);
-    const [result] = await db.execute(
-      "INSERT INTO users (firstname,lastname,email,password) VALUES (?,?,?,?)",
-      [req.body.firstname, req.body.lastname, req.body.email, hash]
-    );
-    if (!result.affectedRows) throw new Error("No Affected Row");
-    res.status(201).json({ message: "User created" });
+    const user = await User.create({
+      firstname,
+      lastname,
+      email,
+      password,
+    });
+    res.status(201).json({ message: "User created", user });
   } catch (e) {
     console.error(e);
     res.status(500).json({ error: "Internal Server Error" });
@@ -73,99 +68,52 @@ exports.signup = async (req, res) => {
 };
 
 /**
- * function connection
+ * function connexion
  */
 exports.login = async (req, res) => {
   if (!req.body.email || !req.body.password)
     return res.status(400).json({ error: "Bad Request" });
+  const { email, password } = req.body;
   try {
-    let [user] = await db.execute(
-      "SELECT id,firstname,lastname,email,password,avatar,is_moderator FROM users WHERE email = ? LIMIT 1",
-      [req.body.email]
-    );
-    if (!user.length) return res.status(401).json({ error: "Unauthorized" });
-    else user = user[0];
-    const valid = await bcrypt.compare(req.body.password, user.password);
-    delete user.password;
-    if (!valid) return res.status(401).json({ error: "Unauthorized" });
-    await db.execute("UPDATE users SET last_logged_in = ? WHERE id = ?", [
-      new Date(),
-      user.id,
-    ]);
-    res.status(200).json({
-      token: jwt.sign({ user }, env.secret, { expiresIn: "24h" }),
-    });
+    const user = await User.scope("admin").findOne({ where: { email } });
+    if (!user) return res.status(401).json({ error: "Unauthorized" });
+    if (!user.isValidPassword(password))
+      return res.status(401).json({ error: "Unauthorized" });
+    user.changed("updatedAt", true);
+    await user.save();
+    res.status(200).json({ token: user.createToken() });
   } catch (e) {
     console.error(e);
     res.status(500).json({ error: "Internal Server Error" });
   }
 };
 
-// exports.list = async (req, res) => {
-//   try {
-//     const [users] = await db.query(
-//       "SELECT id,firstname,lastname,created_at,last_logged_in,email,role_id FROM users"
-//     );
-//     res.status(200).json(users);
-//   } catch (e) {
-//     console.error(e);
-//     res.status(500).json({ error: "Internal Server Error" });
-//   }
-// };
-
-// exports.get = async (req, res) => {
-//   try {
-//     const [users] = await db.execute(
-//       `SELECT
-//         id,
-//         firstname,
-//         lastname,
-//         created_at,
-//         last_logged_in,
-//         email,
-//         role_id
-//       FROM users
-//       WHERE id = ? LIMIT 1`,
-//       [req.params.id]
-//     );
-//     if (!users.length) return res.status(404).json({ error: "User Not Found" });
-//     res.status(200).json(users[0]);
-//   } catch (e) {
-//     console.error(e);
-//     res.status(500).json({ error: "Internal Server Error" });
-//   }
-// };
-
-
 exports.me = async (req, res) => {
- console.log( await isModerator(req.auth.userId));
   if (!req.body.firstname || !req.body.lastname || !req.body.email)
     return res.status(400).json({ error: "Bad Request" });
+  const { firstname, lastname, email, password } = req.body;
   try {
-    const array = [req.body.firstname, req.body.lastname, req.body.email];
-    let query = "UPDATE users SET firstname = ?, lastname = ?, email = ?";
+    const user = await User.findByPk(req.auth.user.id);
+    if (!user) {
+      if (req.file) fs.unlinkSync(req.file.path);
+      return res.status(404).json({ error: "User Not Found" });
+    }
+    user.update({ firstname, lastname, email });
 
     /**
      * verification de la présence d'un mot de passe dans le body
      * si vrai verification de la complexitée et hashage
      * puis on l'ajoute à la requête
      */
-    if (Object.hasOwn(req.body, "password") && req.body.password.length > 0) {
-      let hash;
-      const details = passValidator.validate(req.body.password, {
+    if (password) {
+      const details = passValidator.validate(password, {
         details: true,
       });
       if (details.length)
         return res
           .status(400)
           .json({ error: `Invalid Password: ${details[0].message}` });
-      try {
-        hash = await bcrypt.hash(req.body.password, 10);
-      } catch (e) {
-        return res.status(500).json({ error: "Internal Server Error" });
-      }
-      query += ", password = ?";
-      array.push(hash);
+      user.password = password;
     }
     // "SELECT id,firstname,lastname,email,password,avatar FROM users WHERE email = ? LIMIT 1",
 
@@ -173,54 +121,21 @@ exports.me = async (req, res) => {
      * Verification de la présence d'un fichier dans la requête
      * si vrai on ajoute à notre requête
      */
-    if (req.file) {
-      query += ", avatar = ?";
-      array.push(req.file.filename);
-    }
-    query += " WHERE id = ? LIMIT 1";
-    array.push(req.auth.userId);
-    const [result] = await db.execute(query, array);
-    if (!result.affectedRows) {
-      if (req.file) {
-        fs.unlinkSync(req.file.path);
-      }
-      return res.status(404).json({ error: "User Not Found" });
-    }
-    const user = {
-      firstname: req.body.firstname,
-      lastname: req.body.lastname,
-      email: req.body.email,
-      id: req.auth.userId,
-      avatar: req.file ? req.file.filename : (req.auth.user.avatar || null),
-      is_moderator: await isModerator(req.auth.userId)
-    }
-    if (req.auth.user.avatar && req.file) {
-      fs.unlinkSync(path.join(__dirname,"..", "upload" , req.auth.user.avatar))
-    }
+    if (req.file) user.avatar = req.file.filename;
+
+    // On sauvegarde le modele
+    user.save();
+
+    // Si ancien avatar, on le supprime
+    if (req.auth.user.avatar && req.file)
+      fs.unlinkSync(path.join(__dirname, "..", "upload", req.auth.user.avatar));
 
     res.status(200).json({
-      token: jwt.sign({ user }, env.secret, { expiresIn: "24h" }),
+      token: user.createToken(),
     });
   } catch (e) {
     console.error(e);
-    if (req.file) {
-      fs.unlinkSync(req.file.path);
-    }
-    res.status(500).json({ error: "Internal Server Error" });
-  }
-};
-
-exports.delete = async (req, res) => {
-  try {
-    const [result] = await db.execute(
-      "DELETE FROM users WHERE id = ? LIMIT 1",
-      [req.params.id]
-    );
-    if (!result.affectedRows)
-      return res.status(404).json({ error: "User Not Found" });
-    res.status(200).json({ message: "User Deleted" });
-  } catch (e) {
-    console.error(e);
+    if (req.file) fs.unlinkSync(req.file.path);
     res.status(500).json({ error: "Internal Server Error" });
   }
 };
